requirements.txt
fastapi
uvicorn
ddtrace
requests
opentelemetry-api